+++ RELEASE_NOTES.md
@@ -6,6 +6,8 @@
   - installPlugin skips duplicate JdbiPlugins (according to Object.equals)
   - KotlinSqlObjectPlugin will install forgotten SqlObjectPlugin for you
   - ClasspathSqlLocator allows disabling comment stripping and deprecate static API
+  - Freemarker allows customizing Configuration
+  - FreemarkerSqlLocator should now be created via FreemarkerConfig so it shares config
 - New Beta Features
   - added `register` methods for qualified factories on `Configurable`,
     `ColumnMappers`, and `ArgumentFactories`